Fixed the container creation on LSF hosts where /run/user/$UID doesn't doesnt exist yet
LSF doesn't login the user on the host, so pam doesn't create the
/run/user/$UID directory. Without this directory, the container cant be
invoked with cgroups enabled. Soooo - ive got a trick to enable that by
implementing a pre-exec command to run with the lsf command. eg adding
this switch:
-E "/usr/bin/loginctl enable-linger $USER"

That solves it - hooray!

but its hacky to me, so i've got a ticket open with IBM

ibm.com/mysupport/s/case/500gJ000005GV8DQAW/need-to-initialize-user-login-on-job-submission-but-l-switch-isnt-working