Fixed the LSF job logfile Added a check to verify that the users $HOME exists before submitting a job
- The original LSF logfile used the variable $HOME which is not correct -
thats the inherited variable for the user running the server. Instead,
it needed to use ~{user} for whoever the authenticated user is
- For any new employee who has not yet logged into a linux box, they
  will not yet have a $HOME created and mounted and that results in a
silent failed LSF submission. Added a check to verify that the home
exists before running bsub command, and spitting out a informational
error box for how to fix this if the home doesnt exist