Cleaned up the LSF job submission by adding the "-cwd" and "-eo" switches and also fixing their values
Trying to get the LSF job logfiles to show up by using the resolved
homepath for the authenticated user. Everything looks correct but the
stupid logs still arent showing up. The switches are all good for the
bsub command to there's no harm in adding them, but something is still
missing